Makes get_column_key private to manager.py
The function only gets used in one place, so we can move it there and
mark it private.
